Do not use gelu for fused_linear_act.
Turns out that it is much slower
Also include *.cu file in code snap shot.

Also changed to use int64 for addressing for act_backward.cu. int(int32) may not be enough for large tensors.

+++ alf/ext/act_backward.cu
@@ -43,7 +43,7 @@ __global__ void relu_backward_kernel(const T* grad_output,
                                      const T* input,
                                      T* grad_input,
                                      int n) {
-  int idx = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t idx = blockIdx.x * blockDim.x + threadIdx.x;
   if (idx < n) {
     grad_input[idx] = relu_grad(input[idx], grad_output[idx]);
   }
